Thread an actor map through everything
This is necessary to figure out which object (and therefore class) an
actor is.

This adds a lot of complexity, but I think it's the simplest way to do
this. The alternatives are (a) introducing StateT or (b) using an unsafe
global mutable IORef. The former is way too complex for this library and
the latter is just gross.