Handle vote actors and trailing bits
This is kind of cheating because I don't know why the bits are there. I
don't think they're frame data because trying to parse them fails. And
they're after the last frame. So for the time being I'll just store the
raw data. (I only have one replay with trailing data, so it's not a big
deal.)